fetchChallenges.spec.js
